Create ScrollingGame
For this project I produces a simple scrolling game; where a player tries to avoid a "harm" object and collect a "benefit" object. In the game, my player character (demonstrated as a rocket ship) is positioned near the left edge of the screen and moves up and down based on the user pressing the up and down arrow keys. I have also commented out lines if the user would like to make it a normal space invader game and move the rocket ship side to side, though change in positioning of the objects would have to be made.

I have implemented the collision detection algorithm in order to demonstrate a positive/negative change in score whenever the harm/benefit object comes in contact with the player object. When lives are gone, the game ends and a game over message is displayed.